+++ WTG.Analyzers/Analyzers/Var/VarAnalyzer.cs
@@ -51,7 +51,7 @@ namespace WTG.Analyzers
 
 			var candidate = Visitor.Instance.Visit(context.Node);
 
-			if (candidate == null || candidate.ValueSource.IsKind(FutureSyntaxKinds.DefaultLiteralExpression))
+			if (candidate == null || candidate.ValueSource.IsKind(SyntaxKind.DefaultLiteralExpression))
 			{
 				return;
 			}
