--- WTG.Analyzers/FutureSyntaxKinds.cs
@@ -1,9 +1,0 @@
-using Microsoft.CodeAnalysis.CSharp;
-
-namespace WTG.Analyzers
-{
-	static class FutureSyntaxKinds
-    {
-		public const SyntaxKind DefaultLiteralExpression = (SyntaxKind)8755;
-	}
-}
