+++ appveyor.yml
@@ -1,4 +1,4 @@
-version: 2.0.8.{build}
+version: 2.10.0.{build}
 
 skip_branch_with_pr: true
 os: Visual Studio 2019
@@ -13,7 +13,7 @@ configuration:
   - Release
 
 environment:
-  WTG_ANALYZERS_VERSION: 2.0.8
+  WTG_ANALYZERS_VERSION: 2.10.0
 
 init:
   - git config --global core.autocrlf true
